executor: Move definitions of externals out of wasm_executor module.
